requirements.txt
fastapi
uvicorn[standard]
motor
httpx
python-dotenv
pydantic
whatsapp-webhook-client
requests

